Ensure to specify dir permission
Closes: #3521

FileUtils.mkdir_p is not used with mode: so even though
system_config.dir_permission is specified, it was not used at all.

For command line tools, Fluent::DEFAULT_DIR_PERMISSION
is enough. (maybe no need to follow sticky bit and so on)

For supervisor, changed to follow dir_permission.

Signed-off-by: Kentaro Hayashi <hayashi@clear-code.com>